restful03/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==2.2
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
Pygments==2.7.4
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-django==4.1.0
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
